requirements.txt
gradio==3.27.0
langchain==0.0.151
python-dotenv==1.0.0
